users/requirements.txt
Flask==0.10.1
requests==2.5.1
mysql-connector
